:recycle: refactor: useLocalStorageState, useMockApiState 분리
