--- src/const.ts
@@ -1,11 +1,0 @@
-export const LOCAL_KEYS = Object.freeze({
-  COUPON: "coupon_key",
-  PRODUCT: "product_key",
-  CART_ITEM: "cart_item_key",
-});
-
-export const API = Object.freeze({
-  PRODUCT: "/products",
-  COUPON: "/coupons",
-  CART_ITEM: "/cart-items",
-});
