+++ src/refactoring/hooks/index.ts
@@ -1,6 +1,8 @@
 export * from "./useCart";
 export * from "./useCoupon";
 export * from "./useEditingProduct";
+export * from "./useLocalStorageState";
+export * from "./useMockApiState";
 export * from "./useNewProduct";
 export * from "./useProduct";
 export * from "./useStateByMode";
