+++ src/refactoring/hooks/useCart.ts
@@ -2,14 +2,9 @@ import { useState } from "react";
 import { CartItem, Coupon, Product } from "../../types";
 import { calculateCartTotal, updateCartItemQuantity } from "../models/cart";
 import { useStateByMode } from "./useStateByMode";
-import { API, LOCAL_KEYS } from "../../const";
 
 export const useCart = () => {
-  const [cart, setCart] = useStateByMode<CartItem[]>(
-    LOCAL_KEYS.CART_ITEM,
-    API.CART_ITEM,
-    [],
-  );
+  const [cart, setCart] = useStateByMode<CartItem[]>("CART_ITEM", []);
   const [selectedCoupon, setSelectedCoupon] = useState<Coupon | null>(null);
 
   const addToCart = (product: Product) => {
