+++ src/refactoring/hooks/useCoupon.ts
@@ -1,11 +1,9 @@
-import { API, LOCAL_KEYS } from "../../const.ts";
 import { Coupon } from "../../types.ts";
 import { useStateByMode } from "./useStateByMode.ts";
 
 export const useCoupons = (initialCoupons: Coupon[]) => {
   const [coupons, setCoupons] = useStateByMode<Coupon[]>(
-    LOCAL_KEYS.COUPON,
-    API.COUPON,
+    "COUPON",
     initialCoupons,
   );
