+++ src/refactoring/hooks/useProduct.ts
@@ -1,11 +1,9 @@
 import { Product } from "../../types.ts";
 import { useStateByMode } from "./useStateByMode.ts";
-import { API, LOCAL_KEYS } from "../../const.ts";
 
 export const useProducts = (initialProducts: Product[]) => {
   const [products, setProducts] = useStateByMode<Product[]>(
-    LOCAL_KEYS.PRODUCT,
-    API.PRODUCT,
+    "PRODUCT",
     initialProducts,
   );
