+++ src/refactoring/hooks/useStateByMode.ts
@@ -1,53 +1,21 @@
-import { useEffect, useState } from "react";
-import { localStorageUtil } from "../utils";
-import { apiUtil } from "../utils/apiUtil";
+import { useState } from "react";
+
+import { LOCAL_KEYS, useLocalStorageState } from "./useLocalStorageState";
+import { API, useMockApiState } from "./useMockApiState";
 
 export type EnvMode = "DEFAULT" | "LOCAL_STORAGE" | "API_MOCK";
 
-export const useStateByMode = <T>(
-  key: string,
-  api: string,
-  initialValue: T,
-) => {
+export type StateType = "PRODUCT" | "COUPON" | "CART_ITEM";
+
+export const useStateByMode = <T>(stateType: StateType, initialValue: T) => {
   const envMode = process.env.REACT_APP_MODE as EnvMode;
 
   if (envMode === "LOCAL_STORAGE") {
-    const [state, setState] = useState<T>(
-      () => localStorageUtil.get<T>(key) ?? initialValue,
-    );
-
-    const updateState = async (updater: (prevState: T) => T) => {
-      localStorageUtil.set<T>(key, updater(state));
-      const updatedState = localStorageUtil.get<T>(key);
-
-      if (!updatedState) {
-        throw Error("Updated state is not found.");
-      }
-
-      setState(updatedState);
-    };
-
-    return [state, updateState] as const;
+    return useLocalStorageState<T>(LOCAL_KEYS[stateType], initialValue);
   }
 
   if (envMode === "API_MOCK") {
-    const [state, setState] = useState<T>(() => initialValue);
-
-    useEffect(() => {
-      initState(api);
-    }, []);
-
-    const initState = async (api: string) => {
-      const data = await apiUtil.get<T>(api);
-      setState(data);
-    };
-
-    const updateState = async (updater: (prevState: T) => T) => {
-      const updatedState = await apiUtil.put<T>(api, updater(state));
-      setState(updatedState);
-    };
-
-    return [state, updateState] as const;
+    return useMockApiState<T>(API[stateType], initialValue);
   }
 
   return useState(initialValue);
